app/requirements.txt
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.7
requests==2.25.1

